requirements.txt
python_dotenv
pydantic-ai
generative-ai-hub-sdk[all]
openai
logfire
devtools